Untested rework to remove MusicBrainz. See below for reason.
MusicBrainz's DB is very unreliable, missing a lot of data, not consistent with YouTube. YouTube isn't perfect by any stretch but if we're downloading from YouTube we might as well keep the metadata we get there instead of failing to match to MusicBrainz 50% of the time with no matching MusicBrainz albums/tracks 10% of the time.
I will be testing this rework shortly but first I'm temporarily adding a migrate command in my local copy to migrate my already-downloaded music.